mainrepo/terrarium/dl_repmanager/pyproject.toml
[tool.poetry]
name = "dl-repmanager"
version = "0.0.1"
description = ""
authors = ["Grigorii Statsenko <gstatsenko@yandex-team.ru>"]
packages = [{include = "dl_repmanager"}]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
attrs = ">=22.2.0"
frozendict = ">=2.3.8"
python = ">=3.10, <3.12"
pyyaml = ">=5.3.1"
tomlkit = "==0.11.8"
requests = ">=2.31.0"

[tool.poetry.group.tests.dependencies]
pytest = ">=7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = []

[tool.ruff]
line-length = 140
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I"
]

[tool.black]
line-length = 140
target-version = ['py310']

[tool.poetry.scripts]
dl-repmanager-cli = "dl_repmanager.scripts.repmanager_cli:main"
dl-package-meta = "dl_repmanager.scripts.package_meta_cli:main"

[datalens_ci]
skip_test = true
